PermissionAuthorizationVisitor as separate class
